+++ lib/DataFixtures/Fixture/ProductCategoriesFixture.php
@@ -1,13 +1,13 @@
-<?php namespace Vankosoft\CatalogBundle\DataFixtures\VankosoftExtendedDataFixtures;
+<?php namespace Vankosoft\CatalogBundle\DataFixtures\Fixture;
 
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Vankosoft\ApplicationInstalatorBundle\DataFixtures\AbstractResourceFixture;
 
-final class PaidServiceCategoriesFixture extends AbstractResourceFixture
+final class ProductCategoriesFixture extends AbstractResourceFixture
 {
     public function getName(): string
     {
-        return 'paid_service_categories';
+        return 'product_categories';
     }
     
     protected function configureResourceNode( ArrayNodeDefinition $resourceNode ): void
@@ -19,7 +19,6 @@ final class PaidServiceCategoriesFixture extends AbstractResourceFixture
                 ->scalarNode( 'title' )->end()
                 ->scalarNode( 'description' )->end()
                 ->scalarNode( 'taxonomy_code' )->end()
-                ->scalarNode( 'taxon_code' )->end()
         ;
     }
 }